Updated Mapping Content
Note: Some content on the confluence page did not have an equivalent on the FHIR site. In the interest of preserving the original structure, I left this content out for now. But if it would be helpful to come back and make new pages to make the FHIR site as similar to the confluence page as possible, I am happy to do so!